WIP for User Stories US07192025309PM and US07192025306PM
So training times are saved for each individual piece of gear to a .json file. Also so training times are deleted for each individual piece of gear when said gear is deleted from the list of gear in the Home Screen. Also so the timer window is re-initiated each time instead of re-used. Finally so some naming inconsistencies are improved.