Improved Visuals and Voice Commands
So "Botched" verbal command is replaced with "Eighty-Six". Also so the Home UI buttons to remove, add and nuke gear are visually replaced to words instead of symbols. and window size in the Home UI is adjusted to accommodate for said new labels. Also so the "Performance Tracking" button in the training options has its label replaced to "Graph" for easier understanding from the user. Also so the incorrect dearpygui import and missing initialization for dearpygui is fixed in the Model.py. Finally so the popup text for when there is nothing to graph is more understandable by the user.