WIP US081520251007PM
So that the entered username and password are successfully passed to the Model. So that the username and its corresponding password are stored in a more secure manner via hashing. Finally so that file naming is probed for accuracy with further corrections for file naming in conventions in Model pending.