WIP US081520251007PM
So that a popup is displayed if a user tries to create an account and their username already exists. Also, so input fields are properly deleted after use. Also, so the username and password values are gotten in Model instead of Control. Also, so that the created username is stored in a non-salted hash and its corresponding password is stored as a salted-hash. Finally so the code is cleaned and comments are added.