WIP US081520251007PM
So that a user can initiate a loading sequence for their data upon successful login. Also so that, the header text in the home screen can change based on what user is loaded in. Also so that, a popup happens on successful user creation. Also so that, corresponding popups happen to notify user of their data load sequence. Finally so that, a comment is added to explain the significance of why the log file is created in such a way.