WIP US081520251007PM
So that a suer can only access training graph when they're logged in and they are notified of this. Also so that, a bug is fixed where gear wasn't being added correctly to home screen. Finally so that, popup "labels" are no longer used as messages and unnecessary returns are removed.