WIP US081520251007PM
So that the, user can only train deployment speeds if they are logged in and they are notified of it. Also so that the, Control script does not have control stolen by the Training Options script. Finally so that the Training Options window is utilized better instead of unnecessary bloating in the Track script.