WIP US081520251007PM
So that, a user can logout. Also so that, a bug where data wouldn't visually load properly after login is fixed. Also so that, popups are added for user edge case activity. Also so that, expected user input values are space restricted in ways the user is notified. Finally so that, some redundancy and function naming are cleaned.