WIP US081520251007PM
So that the, user can only logout if they are logged in. Also so that the, flawed logic for the logout function is fixed because previously it would clear gear even if a logout was nonsensical. Also so that, a bug is fixed where an invalid name would showup during failed login attempt of a non-existent user. Finally so that the, input field hints and button labels are improved.