Minor Fixes
So that, dpg items are cleared from memory on logout and account deletion.  Also so that, a message pops up if there are no scores available for a gear item. Also so that, manual "Stop" & "Eighty-Six" UI is hidden upon usage of the stop_training_early function. Finally so that the, code is cleaned.